amount type changed double to bigdecimal

+++ src/main/java/org/project/currencyexchange/model/api/CurrencyLayerResponse.java
@@ -2,10 +2,11 @@ package org.project.currencyexchange.model.api;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 @Data
 public class CurrencyLayerResponse {
     private boolean success;
-    private Map<String, Double> quotes;
+    private Map<String, BigDecimal> quotes;
 }
